+++ classes/conversationHandlers/CreateTrainingSessionConversationHandler.class.js
@@ -6,28 +6,6 @@ CreateTrainingSessionConversationHandler.prototype = {
 	handle: function(from, text, conversation){
 		var that = this;
 		console.log('Handling {'+conversation.topic+'} with {'+from+'}, he said "'+text+'". State is {'+conversation.state+'}');
-		if (conversation.state == 'presenters'){
-			/*if (text.indexOf("me") != -1){*/
-				/*this.bot.say(from, "Cool! I thought so. You're a very good presenter.");*/
-				this.bot.conversationManager.setConversationData(conversation, 'presenter', from, function(){});
-				this.bot.conversationManager.changeConversationState(conversation, 'sessionTitle', function(){
-					that.bot.say(from, "First, what's this session going to be called? Type in the title for the session as you want it to appear for everyone else.");
-				});
-			/*} else if (text){*/
-				/*var presenter = this.bot.persistence.getPlayerByName(text, function(player, err){
-					if(err || !player){
-						that.bot.say(from, "Sorry, I don't know who "+text+ " is. Can you double-check?");
-					} 
-					else {
-						that.bot.say(from, "Great! I'm sure " + player.name + " will do a great job.")
-						that.bot.conversationManager.changeConversationState(conversation, 'sessionTitle', function(){
-							that.bot.say(from, "Now, what's this session going to be called? Type in the title for the session as you want it to appear for everyone else.");
-						});
-					}
-				});				
-				
-			}*/
-		}
 		if (conversation.state == 'sessionTitle'){
 			this.bot.say(from, "Your session's title will be: "+text);
 			this.bot.say(from, "It sounds really interesting, let's keep talking about it!");
@@ -116,7 +94,7 @@ CreateTrainingSessionConversationHandler.prototype = {
 			}
 		}
 		if (conversation.state == 'time'){
-			if(/^\d{1,2}:\d{2}([ap]m)?$/.exec(text)){
+			if (/^\d{1,2}:\d{2}$/.exec(text)) {
 				this.bot.say(from, text + " sounds good to me.");
 				this.bot.conversationManager.setConversationData(conversation, 'time', text, function(){});
 				this.bot.conversationManager.changeConversationState(conversation, "save", function(){