+++ test/all.js
@@ -3,6 +3,7 @@ var managers = require('./managers')
 var handlers = require('./conversationHandlers')
 console.log("[Starting (old) tests]")
 /* DateUtils.class#testFormatYYYYMMDD */
+util.dateUtils.testHasPassed();
 util.dateUtils.testFormat();
 util.dateUtils.testSubtractHours()
 util.dateUtils.testGetCronspec()